Rename components.
