+++ ui/MainScreenDeviceList.js
@@ -1,5 +1,5 @@
+let MainScreenDeviceListItem = require('./MainScreenDeviceListItem');
 let dom = require('dominant');
-let MainScreenPairedDevice = require('./MainScreenPairedDevice');
 
 module.exports = ({devices}) => dom.el('div', {
   class: 'text-gray-700 pt-4 flex flex-col',
@@ -14,7 +14,7 @@ module.exports = ({devices}) => dom.el('div', {
   dom.map(
     () => devices, device => dom.el('div', {
     }, [
-      MainScreenPairedDevice({ device })
+      MainScreenDeviceListItem({ device })
     ]),
   ),
 ]);
