+++ ui/MainScreenDeviceListItem.js
