+++ ui/MainScreenShareList.js
@@ -1,5 +1,5 @@
+let MainScreenShareListItem = require('./MainScreenShareListItem');
 let dom = require('dominant');
-let MainScreenShare = require('./MainScreenShare');
 
 module.exports = ({shares}) => dom.el('div', {
   class: 'my-6 text-gray-700 pt-4 flex flex-col',
@@ -15,7 +15,7 @@ module.exports = ({shares}) => dom.el('div', {
   dom.map(
     () => shares, share => dom.el('div', {
     }, [
-      MainScreenShare({ share })
+      MainScreenShareListItem({ share })
     ]),
   ),
 ])
